mailmind/requirements.txt
# -- Email --
# imaplib, smtplib, and email are Python stdlib modules.

# -- LLM --
openai==1.30.1
httpx==0.27.2

# -- Time parsing and timezone --
dateparser==1.2.0
pytz==2024.1

# -- Google APIs --
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.128.0

# -- Configuration --
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

# -- Testing --
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-mock==3.14.0
